fix(web): move MapIdContext.Provider outside Map component
- Context now wraps entire MapView so children receive mapId
- Memoize children filtering to reduce re-renders
- Fix gestureHandling when only scroll disabled (use cooperative)
