--- src/MapIdContext.web.ts
@@ -1,5 +1,0 @@
-import { createContext, useContext } from 'react';
-
-export const MapIdContext = createContext<string | null>(null);
-
-export const useMapId = () => useContext(MapIdContext);
